+++ packages/std/src/math/decimal.rs
@@ -43,12 +43,27 @@ impl Decimal {
 
     /// Creates a Decimal(value)
     /// This is equivalent to `Decimal::from_atomics(value, 18)` but usable in a const context.
+    ///
+    /// ## Examples
+    ///
+    /// ```
+    /// # use cosmwasm_std::{Uint128, Decimal};
+    /// let atoms = Uint128::new(141_183_460_469_231_731_687_303_715_884_105_727_125);
+    /// let value = Decimal::new(atoms);
+    /// assert_eq!(value.to_string(), "141183460469231731687.303715884105727125");
+    /// ```
+    #[inline]
+    #[must_use]
     pub const fn new(value: Uint128) -> Self {
         Self(value)
     }
 
     /// Creates a Decimal(Uint128(value))
     /// This is equivalent to `Decimal::from_atomics(value, 18)` but usable in a const context.
+    #[deprecated(
+        since = "3.0.0",
+        note = "Use Decimal::new(Uint128::new(value)) instead"
+    )]
     pub const fn raw(value: u128) -> Self {
         Self(Uint128::new(value))
     }
@@ -795,6 +810,7 @@ mod tests {
     }
 
     #[test]
+    #[allow(deprecated)]
     fn decimal_raw() {
         let value = 300u128;
         assert_eq!(Decimal::raw(value).0.u128(), value);
@@ -857,7 +873,7 @@ mod tests {
     fn decimal_try_from_signed_works() {
         assert_eq!(
             Decimal::try_from(SignedDecimal::MAX).unwrap(),
-            Decimal::raw(SignedDecimal::MAX.atomics().i128() as u128)
+            Decimal::new(Uint128::new(SignedDecimal::MAX.atomics().i128() as u128))
         );
         assert_eq!(
             Decimal::try_from(SignedDecimal::zero()).unwrap(),